Move make theme into separate file
